Simplify tool elicitation handler
Refactor MCP tool elicitation handling by extracting client capability
detection and fallback response construction into dedicated public types.
Restore ArgumentBinding semantic helper usage, merge elicitation workflow
coverage into the real specification factory test, and add focused tests for
new public contracts.

Also document helper extraction boundaries in AGENTS.md.
