+++ mcp/bootstrap/src/test/java/org/apache/shardingsphere/mcp/bootstrap/transport/capability/tool/MCPToolClarificationPolicyTest.java
@@ -23,7 +23,10 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -58,5 +61,37 @@ class MCPToolClarificationPolicyTest extends AbstractMCPToolSpecificationFactory
                 WorkflowFieldNames.PLAN_ID, "plan-1",
                 "status", "clarifying",
                 MCPPayloadFieldNames.CLARIFICATION_QUESTIONS, List.of())));
+    }
+    
+    @Test
+    void assertCreateClarificationForm() {
+        Optional<MCPToolClarificationPolicy.ClarificationForm> actual = policy.createClarificationForm(createClarifyingPayload(),
+                createPlanningToolDescriptor("database_gateway_plan_encrypt_rule"));
+        assertTrue(actual.isPresent());
+        assertThat(actual.get().planId(), is("plan-1"));
+        assertThat(actual.get().requestedSchema(), is(createExpectedElicitRequestedSchema()));
+    }
+    
+    @Test
+    void assertCreateClarificationFormWithoutPlanId() {
+        Optional<MCPToolClarificationPolicy.ClarificationForm> actual = policy.createClarificationForm(createClarifyingPayloadWithoutPlanId(),
+                createPlanningToolDescriptor("database_gateway_plan_encrypt_rule"));
+        assertTrue(actual.isEmpty());
+    }
+    
+    @Test
+    void assertCreateClarificationFormWithSensitiveQuestion() {
+        Map<String, Object> payload = createClarifyingPayload(createClarifyingQuestion("custom_properties.access-token", "string", false, "Provide access token."));
+        Optional<MCPToolClarificationPolicy.ClarificationForm> actual = policy.createClarificationForm(payload,
+                createPlanningToolDescriptor("database_gateway_plan_encrypt_rule"));
+        assertTrue(actual.isEmpty());
+    }
+    
+    @Test
+    void assertCreateClarificationFormWithAmbiguousFieldBinding() {
+        Map<String, Object> payload = createClarifyingPayload(createClarifyingQuestion("requires_review", "boolean", false, "Require review?"));
+        Optional<MCPToolClarificationPolicy.ClarificationForm> actual = policy.createClarificationForm(payload,
+                createAmbiguousPlanningToolDescriptor("database_gateway_plan_encrypt_rule"));
+        assertTrue(actual.isEmpty());
     }
 }
